+++ diesel/src/lib.rs
@@ -140,8 +140,10 @@
 //! types provided by `chrono`
 //! - `uuid`: This feature flag enables support for (de)serializing uuid values from the database using types
 //! provided by `uuid`
-//! - `network-address`: This feature flag enables support for (de)serializing IP and Macadresse values from the
-//! database using types provided by `ipnetwork`
+//! - `network-address`: This feature flag enables support for (de)serializing
+//! IP values from the database using types provided by `ipnetwork`.
+//! - `ipnet-address`: This feature flag enables support for (de)serializing IP
+//! values from the database using types provided by `ipnet`.
 //! - `numeric`: This feature flag enables support for (de)serializing numeric values from the database using types
 //! provided by `bigdecimal`
 //! - `r2d2`: This feature flag enables support for the `r2d2` connection pool implementation.