+++ diesel/src/pg/types/mod.rs
@@ -6,6 +6,8 @@ pub(in crate::pg) mod date_and_time;
 #[doc(hidden)]
 pub(in crate::pg) mod floats;
 mod integers;
+#[cfg(feature = "ipnet-address")]
+mod ipnet_address;
 #[cfg(feature = "serde_json")]
 mod json;
 mod mac_addr;
@@ -404,23 +406,24 @@ pub mod sql_types {
     /// Alias for `MacAddr` to be able to use it with `infer_schema`.
     pub type Macaddr = MacAddr;
 
-    /// The [`INET`](https://www.postgresql.org/docs/current/static/datatype-net-types.html) SQL type. This type can only be used with `feature = "network-address"`
+    /// The [`INET`](https://www.postgresql.org/docs/current/static/datatype-net-types.html) SQL type. This type can only be used with `feature = "network-address"` or `feature = "ipnet-address"`.
     ///
     /// ### [`ToSql`] impls
     ///
-    /// - [`ipnetwork::IpNetwork`][IpNetwork]
+    #[cfg_attr(feature = "ipnetwork", doc = " - [`ipnetwork::IpNetwork`][IpNetwork]")]
+    #[cfg_attr(feature = "ipnet", doc = " - [`ipnet::IpNet`][IpNet]")]
+    #[cfg_attr(not(any(feature = "ipnetwork", feature = "ipnet")), doc = "N/A")]
     ///
     /// ### [`FromSql`] impls
     ///
-    /// - [`ipnetwork::IpNetwork`][IpNetwork]
+    #[cfg_attr(feature = "ipnetwork", doc = " - [`ipnetwork::IpNetwork`][IpNetwork]")]
+    #[cfg_attr(feature = "ipnet", doc = " - [`ipnet::IpNet`][IpNet]")]
+    #[cfg_attr(not(any(feature = "ipnetwork", feature = "ipnet")), doc = "N/A")]
     ///
     /// [`ToSql`]: crate::serialize::ToSql
     /// [`FromSql`]: crate::deserialize::FromSql
     #[cfg_attr(feature = "ipnetwork", doc = " [IpNetwork]: ipnetwork::IpNetwork")]
-    #[cfg_attr(
-        not(feature = "ipnetwork"),
-        doc = " [IpNetwork]: https://docs.rs/ipnetwork/*/ipnetwork/enum.IpNetwork.html"
-    )]
+    #[cfg_attr(feature = "ipnet", doc = " [IpNet]: ipnet::IpNet")]
     ///
     /// # Examples
     ///
@@ -434,9 +437,8 @@ pub mod sql_types {
     ///     }
     /// }
     ///
-    /// # #[cfg(feature = "network-address")]
+    /// # #[cfg(any(feature = "network-address", feature = "ipnet-address"))]
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use ipnetwork::IpNetwork;
     ///
     /// #     use diesel::insert_into;
     /// #     use self::clients::dsl::*;
@@ -445,7 +447,8 @@ pub mod sql_types {
     /// #         id SERIAL PRIMARY KEY,
     /// #         ip_address INET NOT NULL
     /// #     )").execute(connection)?;
-    /// let addr = "10.1.9.32/32".parse::<IpNetwork>()?;
+    /// // Parsing "ipnet::IpNet" would also work.
+    /// let addr = "10.1.9.32/32".parse::<ipnetwork::IpNetwork>()?;
     /// let inserted_address = insert_into(clients)
     ///     .values(ip_address.eq(&addr))
     ///     .returning(ip_address)
@@ -454,7 +457,7 @@ pub mod sql_types {
     /// #     Ok(())
     /// # }
     /// #
-    /// # #[cfg(not(feature = "network-address"))]
+    /// # #[cfg(not(any(feature = "network-address", feature = "ipnet-address")))]
     /// # fn main() {}
     /// ```
     #[cfg(feature = "postgres_backend")]
@@ -462,23 +465,24 @@ pub mod sql_types {
     #[diesel(postgres_type(oid = 869, array_oid = 1041))]
     pub struct Inet;
 
-    /// The [`CIDR`](https://www.postgresql.org/docs/postgresql/static/datatype-net-types.html) SQL type. This type can only be used with `feature = "network-address"`
+    /// The [`CIDR`](https://www.postgresql.org/docs/postgresql/static/datatype-net-types.html) SQL type. This type can only be used with `feature = "network-address"` or `feature = "ipnet-address"`.
     ///
     /// ### [`ToSql`] impls
     ///
-    /// - [`ipnetwork::IpNetwork`][IpNetwork]
+    #[cfg_attr(feature = "ipnetwork", doc = " - [`ipnetwork::IpNetwork`][IpNetwork]")]
+    #[cfg_attr(feature = "ipnet", doc = " - [`ipnet::IpNet`][IpNet]")]
+    #[cfg_attr(not(any(feature = "ipnetwork", feature = "ipnet")), doc = "N/A")]
     ///
     /// ### [`FromSql`] impls
     ///
-    /// - [`ipnetwork::IpNetwork`][IpNetwork]
+    #[cfg_attr(feature = "ipnetwork", doc = " - [`ipnetwork::IpNetwork`][IpNetwork]")]
+    #[cfg_attr(feature = "ipnet", doc = " - [`ipnet::IpNet`][IpNet]")]
+    #[cfg_attr(not(any(feature = "ipnetwork", feature = "ipnet")), doc = "N/A")]
     ///
     /// [`ToSql`]: crate::serialize::ToSql
     /// [`FromSql`]: crate::deserialize::FromSql
     #[cfg_attr(feature = "ipnetwork", doc = " [IpNetwork]: ipnetwork::IpNetwork")]
-    #[cfg_attr(
-        not(feature = "ipnetwork"),
-        doc = " [IpNetwork]: https://docs.rs/ipnetwork/*/ipnetwork/enum.IpNetwork.html"
-    )]
+    #[cfg_attr(feature = "ipnet", doc = " [IpNet]: ipnet::IpNet")]
     ///
     /// # Examples
     ///
@@ -492,9 +496,8 @@ pub mod sql_types {
     ///     }
     /// }
     ///
-    /// # #[cfg(feature = "network-address")]
+    /// # #[cfg(any(feature = "network-address", feature = "ipnet-address"))]
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use ipnetwork::IpNetwork;
     ///
     /// #     use diesel::insert_into;
     /// #     use self::clients::dsl::*;
@@ -503,7 +506,8 @@ pub mod sql_types {
     /// #         id SERIAL PRIMARY KEY,
     /// #         ip_address CIDR NOT NULL
     /// #     )").execute(connection)?;
-    /// let addr = "10.1.9.32/32".parse::<IpNetwork>()?;
+    /// // Parsing "ipnet::IpNet" would also work.
+    /// let addr = "10.1.9.32/32".parse::<ipnetwork::IpNetwork>()?;
     /// let inserted_addr = insert_into(clients)
     ///     .values(ip_address.eq(&addr))
     ///     .returning(ip_address)
@@ -511,7 +515,7 @@ pub mod sql_types {
     /// assert_eq!(addr, inserted_addr);
     /// #     Ok(())
     /// # }
-    /// # #[cfg(not(feature = "network-address"))]
+    /// # #[cfg(not(any(feature = "network-address", feature = "ipnet-address")))]
     /// # fn main() {}
     /// ```
     #[cfg(feature = "postgres_backend")]